Added Slot booking pages
I tried putting the TT bat image at the top when the screen size is less but it is getting stretched too much making it ugly. Hence I tried to put it as background image. Now the text is not clearly visible. Hence I had to add a white background-color to green colored text for low mobile size devices.

https://www.awesomescreenshot.com/video/8395442?key=f9ef5f548b8106f5c741c20f8400643b